+++ auth/constants.go
@@ -1,0 +1,6 @@
+package auth
+
+const AUTH_REQUIRED_MSG = "Proxy authentication required.\n"
+const BAD_REQ_MSG = "Bad Request\n"
+const AUTH_TRIGGERED_MSG = "Browser auth triggered!\n"
+const EPOCH_EXPIRE = "Thu, 01 Jan 1970 00:00:01 GMT"
